refactor: split setup delegate modules
